refactor(QueryBuilder): interface 추출 및 패키지 재정리
