+++ src/main/java/persistence/sql/ddl/h2/meta/DataType.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.h2;
+package persistence.sql.ddl.h2.meta;
 
 import java.lang.reflect.Field;
