+++ src/main/java/persistence/sql/ddl/h2/meta/Nullable.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.h2;
+package persistence.sql.ddl.h2.meta;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Id;
