+++ src/main/java/persistence/sql/ddl/h2/meta/TableName.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.h2;
+package persistence.sql.ddl.h2.meta;
 
 import jakarta.persistence.Table;
