+++ src/test/java/persistence/sql/ddl/h2/NullableTest.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Id;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import persistence.sql.ddl.h2.meta.Nullable;
 
 import java.lang.reflect.Field;
